+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Fixes redirection after OAuth login when the user is in all the pages
+- Pass the returnUrl query when open login in mobile version
 
 ## [2.9.2] - 2019-04-30
 ### Changed